Hit.Common.File

Description

A module that provides some operation on files within ExIO monad.

Documentation

secureFileOperation :: IO a -> ExIO a

Executes the given IO operation and returns its result or error in ExIO monad

createEmptyDirectory :: FilePath -> ExIO ()

Creates a new empty directory on a given path

createNewFile

Arguments

:: FilePath
path to directory in which file will be placed

-> String
name of file

-> String
file content

-> ExIO ()

Creates new file

createDirectoryIfNotExist :: FilePath -> ExIO ()

Creates a directory if it does not exist

readWholeFile :: FilePath -> ExIO String

Reads whole file. This uses strict IO

writeWholeFile :: FilePath -> String -> ExIO ()

Writes a content to a file

isDirectory :: FilePath -> ExIO Bool

Checks if a given path indicates a directory

getDirectoryEntries :: FilePath -> ExIO [String]

Gets directory contents

isExistingFile :: FilePath -> ExIO Bool

Checks if file exist

removeExistingFile :: FilePath -> ExIO ()

Removes the given file

splitAndGetRest

Arguments

:: FilePath
first (base) path

-> FilePath
second path

-> [FilePath]
spliitted rest of second path

Removes common part of first path from the second one and return splitted rest of second path

createFileWithParentDirectories

Arguments

:: FilePath
path to file

-> String
file content

-> ExIO ()

Creates a file (if parent directories do not exist it also creates them)

removeIfEmptyDirectory :: FilePath -> ExIO Bool

Removes the directory if it is empty