Hit.Common.List

Description

A module that provides some additional and useful operations on lists.

Documentation

concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]

A concatMap version that operates on external monad

findFirstMatching :: (a -> Bool) -> [a] -> Maybe a

Returns the first element from the list that satisfies a given predicate or Nothing if there is no such element

findOnlyMatching :: (a -> Bool) -> [a] -> Maybe a

Returns the only element from the list that satisfies a given predicate or Nothing if there is no such element

skip :: Int -> [a] -> [a]

Skips a given number of elements and return the rest of the list