Hit.Common.Permissions

Description

A module that provides a function that gives file permissions for Hit.

Documentation

getHitPermissions :: FilePath -> ExIO HitPermissions

Returns permissions for a given argument path to file or directory. Allowed permissions are: 040000 for directories, 100755 for executable files, 100644 for normal files