Hit.Common.Repository

Description

A module that provides some basic paths to important Hit files and directories

Documentation

getRepositoryDirectory :: ExIO FilePath

Gets path to main repository directory. This is a current directory

setRepositoryDirectory :: FilePath -> ExIO ()

Sets path to main repository directory. This is current directory

getHitDirectoryPath :: ExIO FilePath

Gets path to special .hit directory where all Hit information are stored

getPathToObjects :: ExIO FilePath

Gets path to directory where Hit objects are stored

getPathToMergeFile :: ExIO FilePath

Gets path to Merge file.

getPathToConfig :: ExIO FilePath

Gets path to Hit configuration path (.hit/.hitconfig)

getPathToRefs :: ExIO FilePath

Gets path to Hit references directory

getPathToHead :: ExIO FilePath

Gets path to Hit head file