Hit.Objects

Description

A module that defines Hit objects used in repository

Documentation

data HitObjectType

Represents a type of Hit object

Constructors

BlobType

TreeType

CommitType

readHitObjectType :: String -> Maybe HitObjectType

Tries to convert the given String into HitObjectType. Returns Nothing if conversion fails

class HitObject a where

A class for all Hit objects (Blobs, Trees and Commits)

Minimal complete definition

size, objectType, getContent

Methods

size :: a -> Int

Returns size of the object

objectType :: a -> HitObjectType

Returns object type

getContent :: a -> String

Returns object content as String

hashObject :: a -> Hash

Gets Hash associated with the current object

compressObject :: a -> ByteString

Compresses the current object and converts it to a ByteString

size :: HitObject a => a -> Int

Returns size of the object

objectType :: HitObject a => a -> HitObjectType

Returns object type

getContent :: HitObject a => a -> String

Returns object content as String

hashObject :: HitObject a => a -> Hash

Gets Hash associated with the current object

compressObject :: HitObject a => a -> ByteString

Compresses the current object and converts it to a ByteString

data Blob

Represents a base Hit object associated with a single file.

Constructors

Blob

Fields

fileContent :: String

Stores a file content

data DirectoryEntry

Represents an entry in directory (a file or subdirectory)

Constructors

DirectoryEntry

Fields

permissions :: HitPermissions

Gets permissions to the entry

entryName :: String

Gets name of the entry

entryHash :: Hash

Gets Hit hash of the entry

data Tree

Represents Hit object associated with a directory

Constructors

Tree

Fields

entries :: [DirectoryEntry]

Gets list of DirectoryEntry that the object contains

data CommitAuthor

Represents an author of the commit

Constructors

CommitAuthor

Fields

name :: String

Gets the name of the author

email :: Email

Gets an email of the author

data Commit

This is a Hit object that stores information about how repository looks like at the given point in time

Constructors

Commit

Fields

tree :: Hash

Gets associated Tree object hash (for the main repository directory)

message :: String

Gets the message of the commit

author :: CommitAuthor

Gets the author of the commit

committer :: CommitAuthor

Gets the committer of the commit

authorTimestamp :: HitTimestamp

Gets the author timestamp (when commit is created)

committerTimestamp :: HitTimestamp

Gets the committer timestamp

parents :: [Hash]

Gets list of parents for that commit