Hit.Repository.Changes

Description

A module that is responsible for finding changes between various repository versions

Documentation

compareDirectoryTrees

Arguments

:: FilePath
path to directory

-> Tree
first version of directory

-> Tree
second version of directory

-> ExIO [Change]
list of changes

Gets a list of changes between two versions of directory

getRepositoryChanges :: ExIO [Change]

Gets a list of changes made in working directory since last commit

getStatus :: ExIO [Change]

Gets a list of changes made in working directory since last commit

compareTrees

Arguments

:: FilePath
path to directory

-> Tree
first version of directory

-> Tree
second version of directory

-> ExIO [Change]
list of changes

Gets a list of changes between two Tree objects associated with the given directory