Hit.Repository.Checkout

Description

Provides functions that change repository state between branches and commits

Documentation

makeHashCheckout :: Hash -> ExIO ()

Changes repository state to be like in the given commit hash

changeBranch :: Branch -> ExIO ()

Changes the current branch and restores repository state to be exaclty as in the last commit from the given branch

changeToCommit :: Hash -> ExIO ()

Changes repository state to be like in the given commit hash. This sets repository to a deteached head mode