Hit.Repository.Commit

Description

A module that provides way to create commits in repository

Documentation

createCommitWithParents :: String -> [Hash] -> ExIO Commit

Creates a commit object with the given message and list of parents and strores it in a repository

createCommit :: String -> ExIO Commit

Creates a commit with a given message and stores it in a repository (commit parent will be last commit from current branch)