Hit.Repository.Diff

Description

A module that provides method to compare versions of files from various commits

Documentation

getDiffFromCurrentVersion :: FilePath -> ExIO [DiffOperation LineRange]

Gets differences in the given file between versions from last commit and working directory

getDiffBetweenCommits :: FilePath -> Hash -> Hash -> ExIO [DiffOperation LineRange]

Gets differences in the given file between two given commits