Hit.Repository.Directory

Description

A module that exports various usefull functions for operationg on Tree object versions

Documentation

listDirectory :: FilePath -> ExIO [FilePath]

Returns a list of entries in a given directory that are not ignored by Hit

getTree

Arguments

:: FilePath
path to directory

-> Bool
specifies whether created Tree will be saved in repository

-> ExIO Tree

Creates and returns a Tree object for the current directory

findFileInTree :: FilePath -> Tree -> ExIO Blob

Finds a Blob object corresponding with a given file in a given tree.

removeFileFromTree :: FilePath -> ExIO ()

Removes a given file from a repository