Hit.Repository.General.Data

Description

A module that provides some base datatypes used in Hit repository operations

Documentation

data LogEntry

Represents a single entry in commits history

Constructors

LogEntry

Fields

commitHash :: Hash

commitAuthor :: CommitAuthor

commitDate :: String

commitMessage :: String

data MatchingEntry

Represents a base change in repository

Constructors

Matching FilePath DirectoryEntry DirectoryEntry
Given directory entries mathes (are the same)

NewEntry FilePath DirectoryEntry
Given directory entry is new

RemovedEntry FilePath DirectoryEntry
Given directory entry is removed

data Change

Represents a change made on file in repository

Constructors

Modified FilePath
The file is modified

New FilePath
The file is newly added

Removed FilePath
The file is removed

isNew :: Change -> Bool

Checks if the change is associated with new file

isRemoved :: Change -> Bool

Checks if the change is associated with removed file

isModified :: Change -> Bool

Checks if the change is associated with modified file

getPath :: Change -> FilePath

Gets the file path associated with that change

data MergeConflict

Represents a merge conflict

Constructors

RemovedConflict FilePath
There is no file in merged branch version but it is on current branch

ModifiedConflict FilePath
The file has been modified on merged branch