Hit.Repository.General.Ignore

Description

Provides method that list directory, but ignores entries not included in Hit

Documentation

getNotIgnoredDirectoryEntries :: FilePath -> ExIO [FilePath]

Gets content of the given directory, but ignores files and directories not important to Hit