Hit.Repository.General.References

Description

A module that provides functions to manage Hit references

Documentation

getCurrentBranch :: ExIO (Maybe Branch)

Gets the current branch. Returns Nothing if repository is in deteached head mode

isInDeteachedMode :: ExIO Bool

Checks if repository is in deteached head mode

getCommitFromHash :: Hash -> ExIO Commit

Get commit from the given Hash

getTreeFromHash :: Hash -> ExIO Tree

Gets tree from the given Hash

getLastCommitHash :: ExIO Hash

Gets last commit hash

getCurrentBranchVersion :: ExIO Tree

Gets Tree for current branch

getBranchCommitHash :: Branch -> ExIO Hash

Gets last commit for a given branch

writeCommit :: Hash -> ExIO ()

Write a given commit hash to a current branch

writeCommitDeteachedHead :: Hash -> ExIO ()

Writes a given commit hash like a commit in deteached head mode

getFullHash :: Hash -> ExIO Hash

Finds full Hash for the given part if it is unique

getVersion :: Hash -> ExIO Tree

Gets Tree object associated with the given commit Hash

isInMergeState :: ExIO Bool

Gives information whether repository is in merge state (state after merge with conflicts)

getMergeParents :: ExIO [Hash]

Gets future merge commit parents from Merge file

setMergeParents :: [Hash] -> ExIO ()

Sets future merge commit parents to Merge file