Hit.Repository.Reset

Description

A module that provides method for reseting changes made in working directory

Documentation

resetChangesInFile :: FilePath -> ExIO ()

Resets state of a given file to the state from last commit

resetAllChanges :: ExIO ()

Resets all changes made in working directory since last commit