Hit.Store

Description

A module that provides a way to store and restore Hit objects to and from files.

Documentation

getPathToObject :: Hash -> ExIO FilePath

Returns a path for the object with a given hash

storeObject :: HitObject a => a -> ExIO ()

Stores a givent Hit object in a file

getHitObjectType :: FilePath -> ExIO HitObjectType

Restores object type from the given file

restoreBlob :: FilePath -> ExIO Blob

Restores Blob object from the given file

restoreTree :: FilePath -> ExIO Tree

Restores Tree object from the given file

restoreCommit :: FilePath -> ExIO Commit

Restores Commit object from the given file